Replace mapping for visual mode star
I didn't like the readability of the code, or the "feature" that
replaced whitespace with \s\+. Also it was apparently from
https://github.com/bronson/vim-visual-star-search/blob/master/test-patterns

The new one I cobbled together from various visual star implementations
I saw. It uses exact matches for whitespace, and it seems to have better
pattern matching in general. It passes the test patterns at
https://github.com/bronson/vim-visual-star-search/blob/master/test-patterns