Add mapping to change word under cursor
Assigned to Alt-8 for now. It's basically `*` followed by `cgn`, but
without advancing to the next word.  Pressing `.` repeats it for the
next matching word, and skipping is possible with `n` or `N`.  I'd like
to make this work with a visual selection, but it's not working, so I'll
come back to it later perhaps.

Note: this is kind of like :%s/old/new/gc but a little nicer in my
opinion.