Reorganize a few lines
Was trying to figure out Linux virtual consoles's cursor shaping again,
but ended up just moving comments around, and adding quarter sized block
cursor for replacement mode. From a virtual console it works in bash,
and echoing the escape sequence or using readline in vi-mode, but not in
Vim, so I dunno.