Add ttimeout setting
Match setting from Vim's defaults.vim file.

Also, without ttimeout there is still a 1 second delay when pressing
escape in the terminal, so I changed the comments (to match
defaults.vim). It's timeoutlen that affects that, but setting that low
makes it harder to press things like leader key mappings. Anyway, NeoVim
doesn't seem to have that issue in the terminal, so I'm I sort of
thinking of transitioning to use it instead.