Update Vim Help navigation mappings
Switched to using search() so the jumplist and search history don't get
modified. Added *link* search with the 's' key, and moved |pipe| link
searches to 'x'. Also added Tab key to search for all the different
types of links. Switched to using :pop and :tag to move up and down the
tag list. And lastly, added ]] and [[ to search for section divider
lines.