Update mapping for indenting with spaces
Mainly making sure it only works in visual-block mode, but also changed
it to use the spacebar, and the unindent mapping uses the black hole
register now.

Note: if the mapping is used on more than one column, unindenting will
delete the width of the block, where as indenting always inserts 1
space. I need to figure out a fix for that some day.