Change titlestring to not use getcwd()
Apparently the titlebar is updated every time the statusbar is updated,
which is every cursor movement for me (because of row/col numbers). So
having getcwd() in my titlestring ended up being called every cursor
movement, which is really unnecessary. To fix this, I made an autocmd to
update a global variable that keeps track of it only when the DirChanged
event happens.

Note: I found this out after using strace --trace=%file on vim.