Add 'a' option to guioptions
This of course applies to GVim only (under Linux?).

The 'a' option has Vim copy the visual selection to the '*' register,
and I can also access selected (but not copied) text from outside of Vim
with the '*' register as well. Copied text from the clipboard (Ctrl-v
outside Vim) uses the '+' register. It's possible I don't want this
feature, but thought I would try it and see if I run into any issues.

Also added a comment about the 'd' option that asks Vim to use a dark
GTK theme.