Update json ftplugin
Added gq mapping that runs formatprg, and shows any errors in a location
window. Without this fancy mapping, the text you're trying to format
gets replaced by an error message, and then you have to undo it. Note
that most of the code for the FormatJsonPrg was adapted from
https://stackoverflow.com/a/12702366

The above also had some weird interaction with the ALE plugin where it
cleared the location list after switching windows, so I disabled ALE for
json buffers. I'm not using ALE for json linting for fixing anyway.