Update termcap shenanigans
Got the cursor a bit closer to how to want it in the linux tty, which
was previously broken because the termcap stuff for setting cursor
visibility also changed the cursor shape.

Disabled background erase when using screen. Without this fix, the
background outside the text is always black, even on colorschemes that
change the background.

Also did a bit of re-arranging and commenting.