Update vimTodo syntax keywords
Removed the matchadd() function that added a few extra keywords, and
instead moved them to the "after syntax" file, which was doing almost
the same thing.

This change allows the keywords to exist anywhere within the vim
comment, instead of matching at the beginning of a comment. It shouldn't
cause too many problems since most words need to be uppercase, but might
be annoying with the keyword "Note" which I use somewhat often.