Update wombat_mod colorscheme
Switched to using lifepillar's vim-colortemplate plugin to generate it,
which is used to create the official Vim colorschemes.

It should be more complete, and it even passes the check_colors.vim tool
that comes with Vim. I just hope I copied everything correctly from the
old theme.

The template file is also included in colors/templates/