Change terminal alt key support to use termcap
Instead of using mapping, like "map <A-a> \ea" I'm now using `set
<A-a>=\ea` which adds it to Vim's termcap settings. This fixes the long
delay I had after pressing Esc in a terminal. If you use imap or cmap,
it seems to trigger that longer delay, so I replaced it with the other
way of adding alt keys.

Mappings that use shift, like <A-S-c> I converted to <A-C> instead. This
means using the '-S' in the mapping doesn't work. I could fix it in the
loop that generates the mappings, but I prefer to keep it simple.

Also the 'timeout' option was added, even though it's default. Comments
for those timeouts were updated too.