Remove sidescroll option
The default is 0, which horizontally scrolls half a screen and puts the
cursor in the center after hitting the right edge of the screen (e.g.
when using w or l keys). It's a little distracting, but I can see more
of what's to the right compared to when this option was set to 1.