Remove Grep command
Too many issues with escaping. I just want it to act as if I was typing
it in a terminal, which is how the built-in :grep command works. The
problem with :grep though is that it first outputs its results on
screen with a "more" prompt, so it needs to be prefixed with :silent to
prevent that. To do that I used a fancy cabbrev which turns ":grep" into
":silent grep", but only if 'grep' appears at the beginning.

Also changed some grep settings to use :let so I can get rid of the
escaping of spaces.

Removed --smart-case from ripgrep's arguments too since I think I'd
rather search for exact matches, and can add -i to ignore case if
needed.