Modify blank line commands
ReduceBlankLines and DeleteBlankLines basically did similar things, but
only one was trying to keep the view from changing. Using winsaveview()
didn't seem to help, so I just used marks, which work better. Both
commands use these now.

ReduceBlankLines was applying to the whole file, so I added -range, so
that it can be used with visual selections.

Also removed the "alias" RemoveBlankLines because I didn't want to
repeat it now that it's more than one line of code.