Add UpdateHelp command
This command will update Vim's help tags, making it available to the
:help command.

Added because I always forget to update :help when installing/updating a
plugin. This could be done automatically, like it used to be when I used
the pathogen plugin system, but considering how I very rarely add or
even update plugin, manually doing this is preferred.

Note that 11verbose is added for some visual feedback. It will show all
the doc directories being searched within the runtime paths.