Simplify MakeBanner command
It always kind of bothered me how this command allowed a length
argument, but it had to be 7 char less than you wanted. Really there's
no need to make it customizable since I always intend to keep my vimrc
confined to 80 columns, so now it's just hard-coded to that width and
simplified with :normal.