Remove old commented quickfix autocmd
A long time ago when I was using vim on MS Windows I had issues with ^M
appearing in the quickfix window, so I made the autocmd to remove them.
I don't really use Windows anymore, and haven't seen the problem on
Linux, so it should be safe to remove it. It was mainly only being kept
around as a reminder that it's possible to modify the quickfix list.