Change leader key to space
I kind of hate that I used comma as a leader key since it adds a delay
to its original function, which leads to me to not really using it. I'm
hoping to train myself to use , and ; more when using f/F/t/T so I'm
trying <Space> (again) as the leader.