Remove gvimrc reloading from vimrc
It seems $MYGVIMRC isn't expanded when vimrc is first loaded, but it is
expanded when it's in gvimrc. It makes more sense to have the
auto-reloading in gvimrc anyway.

Also removed the second filename for the BufWritePost event since I
don't think it's needed.

Added '| autocmd!` to the augroup line too. I kind of like that... saw
that in habamax's dotfiles, although he uses 'au!'.