Fix arrow navigation within vimrc
In a previous commit I think I added a space before the '|' character
which caused a <Space> to be appended to the <Down> mapping. The result
was that pressing Down arrow and then Up arrow didn't move to the
previous section.

I also didn't like that the search pattern was getting overwritten, so I
used the search() function instead.