Remove unused mapping in vim help
I never really used the mappings I added to move to different
highlighted items and dividers. Also, the <Tab> mapping conflicts with
the Ctrl-i.

One thing I kind of liked though is the demonstration of using the skip
parameter of search() and skipping over the helpExample syntax groups. I
feel like that could be useful in the future, so hopefully I don't
forget about it now that it's gone.