Remove ToggleMenu command
Before learning Vim, I was using Sublime Text where it toggled the menu
visibility when tapping the Alt key. This command (probably one of the
first) was made to do the same thing in GVim. These days I still use
GVim, but I don't use the menu and the script that populates that menu
isn't even loaded, so it does next to nothing.