Tweak colors in linux virtual console
Changed 'NonText' highlight group to cyan on blue. This mainly
highlights the 'precedes' (<) and 'extends' (>) listchars options, which
appear when wrapping is enabled. It's similar to how it's shown with the
'less' pager for me. Also EndOfBuffer had to be modified to a black
background since it was linked to NonText.

Oh yeah, and TODO in comments was changed to just yellow on black since
dark gray backgrounds don't seem to work in linux virtual console.