Change listchars and fillchars options
Replaced some unicode characters with escape codes. The fillchars option
was also removed from the 'non-unicode' path since it's already what's
used for defaults.

Also instead of an if/else I'm setting the unicode options first and
then overwriting them for specific terminals in an if statement. It's
kind of weird, but I want the unicode ones to come first, and I don't
want to negate the conditions.