Add disk2 drive to viminfo ignore list
The Startify plugin uses viminfo to figure out the list of most recent
files and I guess it does a "file exists" check on every file before
displaying it, which results in vim taking about 6 seconds to start
while it waits for my HDD to come out of sleep mode. Very annoying.
Planning to replace the HDD with an SSD drive, so until then, this
'fix' will have to do.