Update colors in myC syntax
I'm still trying to decide if and how I should override the default 'c'
syntax with this 'myC' syntax. I made some attempts but reverted most of
the changes, except these changes to the colors that I wanted to keep.

For the colorscheme, the general idea is to make the text almost one
color as to not be distracted by my normal syntax colors in wombat_mod
colorscheme. Strings are a similar gray used by Normal text, but shifted
to be slightly "orange", same for symbols, but shifted to "cyan", with C
keywords in bold, and comments a darker gray. I could probably do
without symbols and keywords being highlighted, but I'll leave it for
now... I kinda like it.