Add after/{ftplugin,syntax} for xdg desktop files
Looking at .desktop files makes my head explode when it's filled with
localized strings everywhere. This commit changes those localizations to
appear as comments so it's easier to see Name= and Comment= for example.

Word wrap was also turned off, and just for fun I added Up/Down arrow
key mappings to jump to group headers and scroll them to the top.