Tweak linux virtual console palette colors
Linux virtual consoles don't support true color or 256 colors, but you
can modify the 16 colors to any 24-bit color. I used the escape sequence
which begins with ESC]P followed by the hex color number and a 24-bit
color in hex. A better way is to use the `setvtrgb` command, but I can't
think of a simple way to keep it self-contained in my vimrc.

I also changed the `if` statement to look for the TERM env var that ends
in 'linux' so it could also work with 'screen.linux' (the TERM I used in
both screen and tmux). Unfortunately changing the color palette doesn't
work in tmux though, so I'll have to try coming up with a solution that
uses `setvtrgb` or find working escape codes.