Tweak wombat_mod colorscheme
Mainly changed the visual and search-related colors, and I changed the
Function highlight group to purple. With those changes I was able to
delete the color override script in after/colors/.

Removed the feature to enable transparency. I feel like that should be
done outside a colorscheme.

Moved colors/templates/ to colortemplates/ so that the vim-colortemplate
plugin works without me telling it where to output the colorschemes.