Combine cq and cQ operator pending mappings
I often use cq to "change inside quotes", but that just works for double
quotes. There was cQ for single quotes too, but I really don't want to
think about which kind of quoted string I'm in. So, I figured I could
just write a function that takes cursor position on a line and checks
the left and right side, expanding outward, until it encounters either
quote and basically returns i' or i". With that I only need single
operator pending mapping. It's probably not perfect, but it seems to
work well enough so far.