Update mappings in help ftplugin
Originally I was planning to convert key names to commands, but
apparently <C-]> isn't equivalent to <Cmd>tag<CR> like I thought... at
least not for the <CR> mapping. So whatever, I'll just change everything
to <C-]> or <C-t>.

I also added a mapping for the middle mouse button to search for the
visually selected text.