Add mapping to jump to highlighted item in help
This adds a Tab/Shift-Tab mapping to jump to the various highlighted
items in Vim's documentation. I deleted a similar mapping a while back
because I didn't use it much, but recently I've found myself reaching
for it, so of course I had to bring it back. It's not perfect, but it's
a little better than it was previously.